フラグメントを削除。
