+++ app/src/main/res/layout/activity_main.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
+<LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <fragment
-        android:id="@+id/fragmentBookshelf"
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rvMyBookList"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:name="com.example.bookmanager.BookshelfFragment"/>
+        android:layout_height="wrap_content"
+        android:scrollbars="vertical" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
