--- app/src/main/res/layout/fragment_bookshelf.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    tools:context=".BookshelfFragment">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvMyBookList"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:scrollbars="vertical"/>
-
-</LinearLayout>
